[refactor, reorg] reorganize components folder structure, reorg redux containers
