+++ client/src/components/PlayerInfo/PlayerInfoContent.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import PlayerInfoHeader from './PlayerInfoHeader';
-import PlayerInfo from './PlayerInfo';
+import PlayerStats from './PlayerInfoStats';
 
 const PlayerInfoContent = ({ selectedPlayer }) => (
   <div className="playerinfocontent">
     <PlayerInfoHeader player={selectedPlayer} />
-    <PlayerInfo player={selectedPlayer} />
+    <PlayerStats player={selectedPlayer} />
   </div>
 );
