+++ client/src/components/scatterplot/ScatterplotPlayerNameBackground.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styles from '../../styles/scatterplot/ScatterplotPlayerNameBackground.css';
+import styles from '../../styles/Scatterplot/ScatterplotPlayerNameBackground.css';
 
 const ScatterplotPlayerNameBackground = ({ player }) => (
   <h1 className={styles.scatterplotplayernamebackground}>
