+++ client/src/index.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import store from './store/store';
-import AppContainer from './containers/AppContainer';
+import App from './components/App';
 
 ReactDOM.render(
   <Provider store={store}>
-    <AppContainer />
+    <App />
   </Provider>, document.getElementById('root'),
 );
