+++ client/src/styles/App.css
@@ -2,6 +2,5 @@
   border: solid black 1px;
   display: flex;
   width: auto;
-  height: 100%;
   font-family: 'Roboto', sans-serif;
 }
