+++ client/src/styles/scatterplot/Scatterplot.css
@@ -2,7 +2,6 @@
   position: relative;
   border: solid black 1px;
   width: 100%;
-  height: 100%;
 }
 
 .scatterplotsvg {
